Changed safe_malloc and safe_calloc and some cleanup
Reverted from version 1.0 to 0.1
safe_malloc and safe_calloc NEEDS to take a long long now.
I got in trouble for Davidson.
Integer overflow due to 32 bit ints.
Hopefully i never get in trouble for other tensors.
Bit of cleanup of files

+++ include/macros.h
@@ -20,8 +20,8 @@
     ptr = NULL; \
   } while (0)
 
-void* safe_malloc_helper(int s, size_t t, const char *typ, const char *file, int line, const char *func);
+void* safe_malloc_helper(long long s, size_t t, const char *typ, const char *file, int line, const char *func);
 
-void* safe_calloc_helper(int s, size_t t, const char *typ, const char *file, int line, const char *func);
+void* safe_calloc_helper(long long s, size_t t, const char *typ, const char *file, int line, const char *func);
 
 void safe_free_helper(void *ptr, const char *ptrname, const char *file, int line, const char *func);
